Rename the container acceptance tests to di
